msm: footswitch-8x60: Reset DSI domain along with MDP collapse/restore
Some DSI-related logic lives in MDP's collapsible GDHS domain for
8960, 8064 and 8930. Assert the reset to this domain when the reset
of MDP is reset. Use of the reset-only clocks added here assumes the
display driver is making sure the DSI clocks (not managed by the
clock driver) are enabled for the duration of the collapse and
restore sequences so that the resets propagate.

Change-Id: I18638624f7ac92b9cb7228b014df7a187fe974fb
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/devices-8930.c